requirements.txt
django
pillow
psycopg2
dj-database-url
python-dotenv
pytube
assemblyai
requests
pydub